refactors ShareButton to FriendlyButton ++ refactors a tags to buttons
