+++ src/main.js
@@ -4,8 +4,8 @@ import 'focus-visible';
 // Internal Modules
 import initResponsiveVideo from './components/modular-content/responsive-video/responsive-video';
 import initToggle from './components/toggle/toggle';
-import initShareButton from './components/share-button/share-button';
+import initFriendlyButton from './components/friendly-button/friendly-button';
 
 initResponsiveVideo();
 initToggle();
-initShareButton();
+initFriendlyButton();
