+++ src/main.scss
@@ -40,7 +40,7 @@
 @import 'components/button/button';
 @import 'components/search/search';
 @import 'components/download-list/download-list';
-@import 'components/share-button/share-button';
+@import 'components/friendly-button/friendly-button';
 @import 'components/loader/loader';
 @import 'components/modular-content/modular-content';
 @import 'components/modular-content/rich-text/rich-text';
